requirements.txt
boto3==1.34.54
Flask==3.0.2
pika==1.3.2
psycopg2-binary==2.9.9
Flask-Mail
